Control.Workflow.Configuration

Description

Helpers for application initialization

Documentation

once :: (Typeable a, Serialize a, MonadIO m) => m a -> Workflow m a

to execute one computation once . It executes at the first run only

ever :: (Typeable a, Serialize a, MonadIO m) => IO a -> Workflow m a

to execute a computation every time it is invoked. A synonimous of unsafeIOtoWF

runConfiguration :: (Monad m, MonadIO m, MonadCatch m) => String -> Workflow m a -> m a

executes a computation with once and ever statements a synonym of exec1nc