Control.Workflow.Stat

Documentation

data WF s m l

Constructors

WF

Fields

st :: s -> m (s, l)

data Stat

Constructors

Running (Map String (String, Maybe ThreadId))

Stat

Fields

self :: DBRef Stat

wfName :: String

state :: Int

recover :: Bool

timeout :: Maybe Integer

lastActive :: Integer

context :: (Context, ByteString)

references :: [(Int, (IDynamic, Bool))]

versions :: [IDynamic]

stat0 :: Stat

statPrefix1 :: [Char]

statPrefix :: [Char]

header :: Stat -> STW ()

getHeader :: STR (String, Int, Maybe Integer, Integer, [t], (Context, ByteString))

lenLen :: Int64

keyWF :: Indexable a => String -> a -> String

Return the unique name of a workflow with a parameter (executed with exec or start)

data WFRef a

Constructors

WFRef !Int !(DBRef Stat)

pathWFlows :: [Char]

stFName :: IResource a => a -> [Char]

fr :: Key -> IO (Maybe ByteString)

safe :: FilePath -> (Handle -> IO a) -> IO a

hReadFile :: Handle -> IO ByteString

readHeader :: ByteString -> Handle -> IO (Maybe (String, Int, Maybe Integer, Integer, [t], (Context, ByteString)))

keyRunning :: String

showHistory :: Stat -> ByteString

show the state changes along the workflow, that is, all the intermediate results

showp1 :: IDynamic -> STW ()

wFRefStr :: [Char]